Add auth blueprint and migrate Discord OAuth
