+++ templates/public/events.html
@@ -18,7 +18,7 @@
           <th>📝 {{ t("Titel") }}</th>
           <th>📄 {{ t("Beschreibung") }}</th>
           <th>🗓️ {{ t("Datum") }}</th>
-          {% if session.get("user") %}
+          {% if session.get("discord_user") %}
             <th>⚙️ {{ t("Aktionen") }}</th>
           {% endif %}
         </tr>
@@ -30,7 +30,7 @@
             <td>{{ event.title }}</td>
             <td>{{ event.description[:80] }}{% if event.description|length > 80 %}...{% endif %}</td>
             <td>{{ event.event_time }}</td>
-            {% if session.get("user") %}
+            {% if session.get("discord_user") %}
               <td>
                 <form action="{{ url_for('public.join_event', event_id=event.id) }}" method="POST" style="display:inline;">
                   <button type="submit" class="btn btn-small">🎟️ {{ t("Beitreten") }}</button>
