+++ templates/public/events_list.html
@@ -16,7 +16,7 @@
           <th>📝 {{ t("Titel") }}</th>
           <th>📄 {{ t("Beschreibung") }}</th>
           <th>🕒 {{ t("Datum & Uhrzeit") }}</th>
-          {% if session.get("user") %}
+          {% if session.get("discord_user") %}
             <th>🎟️ {{ t("Aktion") }}</th>
           {% endif %}
         </tr>
@@ -28,7 +28,7 @@
               <td>{{ event.title }}</td>
               <td>{{ event.description[:80] }}{% if event.description|length > 80 %}…{% endif %}</td>
               <td>{{ event.event_time }}</td>
-              {% if session.get("user") %}
+              {% if session.get("discord_user") %}
                 <td>
                   <form action="{{ url_for('public.join_event', event_id=event.id) }}" method="POST" style="display:inline;">
                     <button type="submit" class="btn btn-small">➕ {{ t("Beitreten") }}</button>
