+++ templates/public/index.html
@@ -8,6 +8,6 @@
   <div style="text-align: center; margin-top: 2em;">
     <h1>🐺 {{ t('welcome_message') }}</h1>
     <p>{{ t('portal_intro') }}</p>
-    <a class="btn" href="{{ url_for('public.discord_login') }}">{{ t('login_with_discord') }}</a>
+    <a class="btn" href="{{ url_for('auth.login') }}">{{ t('login_with_discord') }}</a>
   </div>
 {% endblock %}
