+++ templates/public/login.html
@@ -7,6 +7,6 @@
 {% block content %}
   <div style="text-align: center; margin-top: 3em;">
     <h2>{{ t("login_required") }}</h2>
-    <a class="btn" href="{{ url_for('public.discord_login') }}">{{ t("login_with_discord") }}</a>
+    <a class="btn" href="{{ url_for('auth.login') }}">{{ t("login_with_discord") }}</a>
   </div>
 {% endblock %}
