+++ web/auth/decorators.py
@@ -17,9 +17,9 @@ def login_required(view_func):
 
     @wraps(view_func)
     def wrapper(*args, **kwargs):
-        if "user" not in session:
+        if "discord_user" not in session:
             flash(t("login_required", default="Login required."), "warning")
-            return redirect(url_for("public.login"))
+            return redirect(url_for("auth.login"))
         return view_func(*args, **kwargs)
 
     return wrapper
@@ -30,9 +30,9 @@ def r3_required(view_func):
 
     @wraps(view_func)
     def wrapper(*args, **kwargs):
-        if session.get("user", {}).get("role_level") not in ["R3", "R4", "ADMIN"]:
+        if session.get("discord_user", {}).get("role_level") not in ["R3", "R4", "ADMIN"]:
             flash(t("member_only", default="Members only."))
-            return redirect(url_for("public.login"))
+            return redirect(url_for("auth.login"))
         return view_func(*args, **kwargs)
 
     return wrapper
@@ -43,9 +43,9 @@ def r4_required(view_func):
 
     @wraps(view_func)
     def wrapper(*args, **kwargs):
-        if session.get("user", {}).get("role_level") not in ["R4", "ADMIN"]:
+        if session.get("discord_user", {}).get("role_level") not in ["R4", "ADMIN"]:
             flash(t("admin_only", default="Admins only."))
-            return redirect(url_for("public.login"))
+            return redirect(url_for("auth.login"))
         return view_func(*args, **kwargs)
 
     return wrapper
@@ -56,9 +56,9 @@ def admin_required(view_func):
 
     @wraps(view_func)
     def wrapper(*args, **kwargs):
-        if session.get("user", {}).get("role_level") != "ADMIN":
+        if session.get("discord_user", {}).get("role_level") != "ADMIN":
             flash(t("superuser_only", default="Superuser only."))
-            return redirect(url_for("public.login"))
+            return redirect(url_for("auth.login"))
         return view_func(*args, **kwargs)
 
     return wrapper
